Update AssignmentDriver.java
This is the AssignmentDriver configured to be exported as a JAR file. 
It can be run in Hadoop like this:
hadoop jar Assignment2.jar AssignmentDriver /user/cloudera/input/ /user/cloudera/output/
This presupposes that /user/cloudera/input/ has been made like this:

hdfs dfs -mkdir /user/cloudera/input/

And that the data has been loaded into it like this:

hdfs dfs -copyFromLocal ~/Desktop/20-newsgroups/ /user/cloudera/input/ 

(It will take a while for the data to load, and it will complain the whole time, but it seems to work so far. I made a copy of 20-newsgroups to the Desktop for this to work.)